backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mangacolor-g-backend"
version = "1.0.0"
description = "Backend service for MangaColor-G manga colorization tool"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "MangaColor-G Team"}
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
    "PyMuPDF>=1.23.0",
    "pdfplumber>=0.10.0",
    "Pillow>=10.2.0",
    "opencv-python-headless>=4.9.0",
    "google-genai>=0.1.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.0",
    "websockets>=12.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

